webhooks: Move init code into webhooks package
